requirements.txt
Django==1.11.6
django-pb-model==0.1.3
djangorestframework==3.7.0
protobuf==3.4.0
pytz==2017.2
six==1.11.0
